# Bespoke vec methods