--- a/gtsam/geometry/Gal3.cpp
+++ b/gtsam/geometry/Gal3.cpp
@@ -173,7 +173,24 @@ Matrix5 Gal3::matrix() const {
 }
 
 //------------------------------------------------------------------------------
-
+Gal3::Vector25 Gal3::vec(OptionalJacobian<25, 10> H) const {
+    const Matrix5 T = this->matrix();
+    if (H) {
+        H->setZero();
+        auto R = T.block<3, 3>(0, 0);
+        H->block<3, 1>(0, 6 + 1) = -R.col(2);
+        H->block<3, 1>(0, 6 + 2) = R.col(1);
+        H->block<3, 1>(5, 6 + 0) = R.col(2);
+        H->block<3, 1>(5, 6 + 2) = -R.col(0);
+        H->block<3, 1>(10, 6 + 0) = -R.col(1);
+        H->block<3, 1>(10, 6 + 1) = R.col(0);
+        H->block<3, 3>(15, 3) = R;
+        H->block<3, 3>(20, 0) = R;
+        H->block<3, 1>(20, 9) = T.block<3, 1>(0, 3);
+        (*H)(23, 9) = 1;
+    }
+    return Eigen::Map<const Vector25>(T.data());
+}
 
 //------------------------------------------------------------------------------
 // Stream operator