--- a/gtsam/geometry/Pose3.h
+++ b/gtsam/geometry/Pose3.h
@@ -47,6 +47,7 @@ private:
   Point3 t_; ///< Translation gPp, from global origin to pose frame origin
 
 public:
+  using Vector16 = Eigen::Matrix<double, 16, 1>;
 
   /// @name Standard Constructors
   /// @{
@@ -305,6 +306,9 @@ public:
   /** convert to 4*4 matrix */
   Matrix4 matrix() const;
 
+  /// Return vectorized SE(3) matrix in column order.
+  Vector16 vec(OptionalJacobian<16, 6> H = {}) const;
+
   /** 
     * Assuming self == wTa, takes a pose aTb in local coordinates 
     * and transforms it to world coordinates wTb = wTa * aTb.
@@ -382,9 +386,6 @@ public:
    */
   Pose3 slerp(double t, const Pose3& other, OptionalJacobian<6, 6> Hx = {},
                                              OptionalJacobian<6, 6> Hy = {}) const;
-
-  /// Return vectorized SE(3) matrix in column order.
-  Vector vec(OptionalJacobian<16, 6> H = {}) const;
 
   /// Output stream operator
   GTSAM_EXPORT