… gtsam/geometry/SO4.cpp
… gtsam/geometry/SOn.h
… gtsam/geometry/tests/testPose3.cpp
… gtsam/geometry/tests/testSO4.cpp
… gtsam/navigation/NavState.cpp
… gtsam/navigation/NavState.h
… gtsam/navigation/tests/testNavState.cpp
… timing/timeFrobeniusFactor.cpp