Added custom vecs for most types, which are faster. Note, timing is with branch in #2191.

| Factor    | Matrix Size | Before (ns) | After (ns) | % Improvement |
|-----------|-------------|-------------|------------|----------------|
| Rot2      | 4x1         | 41          | 24         | 41%            |
| SO3       | 9x3         | 60          | 41         | 31%            |
| Rot3      | 9x3         | 65          | 45         | 30%            |
| Pose2     | 9x3         | 97          | 44         | 54%            |
| SO4       | 16x6         | 729         | 187        | 74%            |
| Pose3     | 16x6         | 414         | 204        | 50%            |
| NavState  | 25x9         | —           | 375        | —              |
| Gal3      | 25x10         | 635         | 427        | 32%            |

Still mostly dominated by  matrix size. ~Note Rot2 and Pose2 still use generic vec~